refactor(front): changes App to app
